refactor: re-use app bar across screens
