+++ lib/features/wager_created/wager_created_screen.dart
@@ -6,7 +6,12 @@ class WagerCreatedScreen extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     return Scaffold(
-      appBar: !context.isMobile ? WagerCreatedAppBar() : null,
+      appBar: !context.isMobile ? BaseAppbar(
+          context: context,
+          title: '',
+          userName: '@noyi_24',
+          showBackButton: false,
+      ) : null,
       backgroundColor: context.primaryBackgroundColor,
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
       body: SafeArea(
